refactor wip: split content into columns in LongFormPage
allows putting header info (author, publication date, etc...) into the content before splitting into column
